+++ modules/defaults/graphics/src/types.rs
@@ -2,14 +2,14 @@ use interstice_sdk::*;
 use std::str::FromStr;
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct Vec2 {
     pub x: f32,
     pub y: f32,
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct Vec3 {
     pub x: f32,
     pub y: f32,
@@ -17,7 +17,7 @@ pub struct Vec3 {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct Color {
     pub r: f32,
     pub g: f32,
@@ -26,16 +26,16 @@ pub struct Color {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct Rect {
-    pub x: f32,
-    pub y: f32,
-    pub w: f32,
-    pub h: f32,
+    pub x: u32,
+    pub y: u32,
+    pub w: u32,
+    pub h: u32,
 }
 
 #[interstice_type]
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, PartialEq)]
 pub enum Draw2DCommandType {
     Circle,
     Circles,
@@ -47,23 +47,23 @@ pub enum Draw2DCommandType {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct RectCommand {
     pub rect: Rect,
     pub color: Color,
     pub filled: bool,
-    pub stroke_width: f32,
+    pub stroke_width: u32,
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct ResourceAddress {
     pub owner_node_id: String,
     pub local_id: String,
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct CircleCommand {
     pub center: Vec2,
     pub radius: f32,
@@ -73,17 +73,17 @@ pub struct CircleCommand {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct PolylineCommand {
     pub points: Vec<Vec2>,
     pub color: Color,
-    pub width: f32,
+    pub width: u32,
     pub closed: bool,
     pub filled: bool,
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct ImageCommand {
     pub texture: ResourceAddress,
     pub rect: Rect,
@@ -91,7 +91,7 @@ pub struct ImageCommand {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct TextCommand {
     pub content: String,
     pub position: Vec2,
@@ -101,7 +101,7 @@ pub struct TextCommand {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct MeshDrawCommand {
     pub mesh: ResourceAddress,
     pub pipeline: ResourceAddress,
@@ -109,7 +109,7 @@ pub struct MeshDrawCommand {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct MeshVertex {
     pub position: Vec3,
     pub color: Color,
@@ -117,14 +117,14 @@ pub struct MeshVertex {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct MeshDescriptor {
     pub vertices: Vec<MeshVertex>,
     pub indices: Option<Vec<u32>>,
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct PipelineDescriptorInput {
     pub label: Option<String>,
     pub shader_source: String,
@@ -133,13 +133,13 @@ pub struct PipelineDescriptorInput {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct BindGroupDescriptorInput {
     pub label: Option<String>,
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct TextureUsageFlags {
     pub copy_src: bool,
     pub copy_dst: bool,
@@ -149,7 +149,7 @@ pub struct TextureUsageFlags {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct TextureDescriptorInput {
     pub width: u32,
     pub height: u32,
@@ -160,14 +160,14 @@ pub struct TextureDescriptorInput {
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct RenderPassSubmission {
     pub layer: String,
     pub debug_label: Option<String>,
 }
 
 #[interstice_type]
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct ComputeSubmission {
     pub debug_label: Option<String>,
 }
